refactor: RacingGame OutputView 사용 제거.
